move some stuff to src/gtc
