+++ src/gt4py/backend/gtc_backend/defir_to_gtir.py
@@ -1,7 +1,6 @@
 from types import MappingProxyType
 from typing import ClassVar, Dict, List, Mapping
 
-from gt4py.backend.gtc_backend import common, gtir
 from gt4py.ir import IRNodeVisitor
 from gt4py.ir.nodes import (
     Assign,
@@ -18,6 +17,7 @@ from gt4py.ir.nodes import (
     ScalarLiteral,
     StencilDefinition,
 )
+from gtc import common, gtir
 
 
 def transform_offset(offset: Dict[str, int]) -> gtir.CartesianOffset:
