+++ src/gtc/common.py
