+++ src/gtc/gtir.py
