+++ src/gtc/passes/__init__.py
@@ -1,0 +1,1 @@
+from .fields_metadata_pass import FieldsMetadataPass
