+++ tests/test_unittest/test_gtc_backend/test_naive_python_gen.py
@@ -2,8 +2,8 @@ import ast
 
 import pytest
 
-from gt4py.backend.gtc_backend.common import BinaryOperator, LoopOrder
-from gt4py.backend.gtc_backend.gtir import (
+from gtc.common import BinaryOperator, LoopOrder
+from gtc.gtir import (
     AssignStmt,
     AxisBound,
     BinaryOp,
@@ -13,7 +13,7 @@ from gt4py.backend.gtc_backend.gtir import (
     VerticalInterval,
     VerticalLoop,
 )
-from gt4py.backend.gtc_backend.python_naive_codegen import PythonNaiveCodegen
+from gtc.python_naive_codegen import PythonNaiveCodegen
 
 
 def ast_parse(arg):
