+++ tests/test_unittest/test_gtc_backend/test_stencil_module_ast.py
@@ -5,7 +5,6 @@ import black
 import numpy
 import pytest
 
-from gt4py.backend.gtc_backend import gtir
 from gt4py.backend.gtc_backend.stencil_module_builder import (
     FieldInfoBuilder,
     StencilClassBuilder,
@@ -15,6 +14,7 @@ from gt4py.backend.gtc_backend.stencil_module_builder import (
 )
 from gt4py.stencil_object import StencilObject
 from gt4py.utils import make_module_from_file
+from gtc import gtir
 
 
 def test_parse_snippet():
