+++ tests/test_unittest/test_gtc_backend/test_stencil_object_snippet_generators.py
@@ -4,8 +4,15 @@ from typing import Iterator
 
 import pytest
 
-from gt4py.backend.gtc_backend.common import DataType, LoopOrder
-from gt4py.backend.gtc_backend.gtir import (
+from gt4py.backend.gtc_backend.stencil_object_snippet_generators import (
+    ComputationCallGenerator,
+    DomainInfoGenerator,
+    FieldInfoGenerator,
+    ParameterInfoGenerator,
+    RunBodyGenerator,
+)
+from gtc.common import DataType, LoopOrder
+from gtc.gtir import (
     AccessKind,
     AssignStmt,
     AxisBound,
@@ -20,13 +27,6 @@ from gt4py.backend.gtc_backend.gtir import (
     Stencil,
     VerticalInterval,
     VerticalLoop,
-)
-from gt4py.backend.gtc_backend.stencil_object_snippet_generators import (
-    ComputationCallGenerator,
-    DomainInfoGenerator,
-    FieldInfoGenerator,
-    ParameterInfoGenerator,
-    RunBodyGenerator,
 )
